+++ src/seed/seed.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { SeedService } from './seed.service';
+
+@Module({
+  providers: [SeedService]
+})
+export class SeedModule {}
